+++ packages/web3-eth-ens/CHANGELOG.md
@@ -154,3 +154,6 @@ Documentation:
 - Added function getText and getName in ENS and resolver classes (#6914)
 
 ## [Unreleased]
+
+### Added
+-   `defaultReturnFormat` was added to all methods that have `ReturnType` param. (#6947)